refactor: 네이밍 컨벤션 수정
